checkpoint #2 -> refactoring services and change relation between Resident and Apartment

+++ HarmonyHomeNet-Logic/src/main/java/bwp/pwr/daniel/rysz/harmonyhomenetlogic/mainLogicEntitys/basment/service/BasementServiceImp.java
@@ -1,12 +1,13 @@
 package bwp.pwr.daniel.rysz.harmonyhomenetlogic.mainLogicEntitys.basment.service;
 
+import bwp.pwr.daniel.rysz.harmonyhomenetlogic.exeptions.customErrors.BasementNotFoundException;
 import bwp.pwr.daniel.rysz.harmonyhomenetlogic.mainLogicEntitys.basment.entity.Basement;
 import bwp.pwr.daniel.rysz.harmonyhomenetlogic.mainLogicEntitys.basment.repository.BasementRepository;
+import bwp.pwr.daniel.rysz.harmonyhomenetlogic.utils.response.buildingStaff.BasementResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -16,27 +17,51 @@ public class BasementServiceImp implements BasementService {
 
 
     @Override
-    public void deleteById(UUID id) {
-        basementRepository.deleteById(id);
+    public void deleteById(UUID id) throws BasementNotFoundException {
+        if(basementRepository.existsById(id)) basementRepository.deleteById(id);
+        else throw new BasementNotFoundException("wrong basement id");
     }
 
     @Override
-    public Optional<Basement> findById(UUID id) {
-        return basementRepository.findById(id);
+    public BasementResponse findById(UUID id) throws BasementNotFoundException {
+        return basementRepository.findById(id)
+                .map(basement -> BasementResponse.builder()
+                        .id(basement.getId())
+                        .basementNumber(basement.getBasementNumber())
+                        .area(basement.getArea())
+                        .build()
+                ).orElseThrow(() -> new BasementNotFoundException("wrong basement id"));
     }
 
     @Override
-    public Optional<Basement> findByBasementNumber(int basementNumber) {
-        return basementRepository.findByBasementNumber(basementNumber);
+    public BasementResponse findByBasementNumber(int basementNumber) throws BasementNotFoundException {
+        return basementRepository.findByBasementNumber(basementNumber)
+                .map(basement -> BasementResponse.builder()
+                        .id(basement.getId())
+                        .basementNumber(basement.getBasementNumber())
+                        .area(basement.getArea())
+                        .build()
+                ).orElseThrow(() -> new BasementNotFoundException("wrong basement number"));
     }
 
     @Override
-    public List<Basement> findAll() {
-        return basementRepository.findAll();
+    public List<BasementResponse> findAll() {
+        return basementRepository.findAll().stream()
+                .map(basement -> BasementResponse.builder()
+                        .id(basement.getId())
+                        .basementNumber(basement.getBasementNumber())
+                        .area(basement.getArea())
+                        .build()
+                ).toList();
     }
 
     @Override
-    public void save(Basement basement) {
+    public BasementResponse save(Basement basement) {
         basementRepository.save(basement);
+        return BasementResponse.builder()
+                .id(basement.getId())
+                .basementNumber(basement.getBasementNumber())
+                .area(basement.getArea())
+                .build();
     }
 }
