+++ HarmonyHomeNet-Logic/src/main/java/bwp/pwr/daniel/rysz/harmonyhomenetlogic/mainLogicEntitys/parkingSpace/service/ParkingSpaceServiceImp.java
@@ -1,12 +1,13 @@
 package bwp.pwr.daniel.rysz.harmonyhomenetlogic.mainLogicEntitys.parkingSpace.service;
 
+import bwp.pwr.daniel.rysz.harmonyhomenetlogic.exeptions.customErrors.ParkingSpaceNotFoundException;
 import bwp.pwr.daniel.rysz.harmonyhomenetlogic.mainLogicEntitys.parkingSpace.entity.ParkingSpace;
 import bwp.pwr.daniel.rysz.harmonyhomenetlogic.mainLogicEntitys.parkingSpace.repository.ParkingSpaceRepository;
+import bwp.pwr.daniel.rysz.harmonyhomenetlogic.utils.response.buildingStaff.ParkingSpaceResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -17,27 +18,35 @@ public class ParkingSpaceServiceImp implements ParkingSpaceService {
 
 
     @Override
-    public void deleteById(UUID id) {
-        parkingSpaceRepository.deleteById(id);
+    public void deleteById(UUID id) throws ParkingSpaceNotFoundException {
+        if (parkingSpaceRepository.existsById(id)) parkingSpaceRepository.deleteById(id);
+        else throw new ParkingSpaceNotFoundException("wrong parking space id");
+
     }
 
     @Override
-    public Optional<ParkingSpace> findById(UUID id) {
-        return parkingSpaceRepository.findById(id);
+    public ParkingSpace findById(UUID id) throws ParkingSpaceNotFoundException {
+        return parkingSpaceRepository.findById(id)
+                .orElseThrow(() -> new ParkingSpaceNotFoundException("wrong parking space id"));
     }
 
     @Override
-    public Optional<ParkingSpace> findByNumber(int parkingSpaceNumber) {
-        return parkingSpaceRepository.findByNumber(parkingSpaceNumber);
+    public List<ParkingSpaceResponse> findAll() {
+        return parkingSpaceRepository.findAll().stream()
+                .map(parkingSpace -> ParkingSpaceResponse.builder()
+                        .id(parkingSpace.getId())
+                        .number(parkingSpace.getNumber())
+                        .build()
+                ).toList();
     }
 
     @Override
-    public List<ParkingSpace> findAll() {
-        return parkingSpaceRepository.findAll();
-    }
-
-    @Override
-    public void save(ParkingSpace parkingSpace) {
+    public ParkingSpaceResponse save(ParkingSpace parkingSpace) {
         parkingSpaceRepository.save(parkingSpace);
+
+        return ParkingSpaceResponse.builder()
+                .id(parkingSpace.getId())
+                .number(parkingSpace.getNumber())
+                .build();
     }
 }
